Refine the dygraph ptq and the module of calculating KL threshold (#33898)
* refine ptq according comments
* reuse the module to calculate kl threshold
